+++ lib/src/dialogs/macos_alert_dialog.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:flutter/physics.dart';
 import 'package:macos_ui/macos_ui.dart';
 import 'package:macos_ui/src/library.dart';
@@ -227,12 +226,13 @@ class MacosAlertDialog extends StatelessWidget {
                   ),
                 ],
               ],
-              const SizedBox(height: 16),
-              if (suppress != null)
+              if (suppress != null) ...[
+                const SizedBox(height: 16),
                 DefaultTextStyle(
                   style: MacosTheme.of(context).typography.headline,
                   child: suppress!,
                 ),
+              ],
               const SizedBox(height: 16),
             ],
           ),