+++ test/buttons/icon_button_test.dart
@@ -85,8 +85,10 @@ void main() {
       [
         'backgroundColor: null',
         'disabledColor: null',
+        'hoverColor: null',
         'pressedOpacity: 0.4',
         'alignment: Alignment.center',
+        'padding: null',
         'semanticLabel: null',
       ],
     );