+++ test/theme/icon_button_theme_test.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:macos_ui/macos_ui.dart';
 import 'package:macos_ui/src/library.dart';
@@ -50,9 +49,11 @@ void main() {
       [
         'backgroundColor: null',
         'disabledColor: null',
+        'hoverColor: null',
         'shape: null',
         'borderRadius: null',
         'boxConstraints: null',
+        'padding: null',
       ],
     );
   });